Rename miner -> proposer

+++ config/chainfile.go
@@ -9,7 +9,7 @@ import (
 // ChainFile represents the on-disk chain file used to initialize the chain.
 type ChainFile struct {
 	Validators         []primitives.ValidatorInitialization `json:"validators"`
-	GenesisTime        int64                                `json:"genesis_time"`
+	GenesisTime        uint64                                `json:"genesis_time"`
 	InitialConnections []string                             `json:"initial_connections"`
 	PremineAddress     string                               `json:"premine_address"`
 }
@@ -19,7 +19,7 @@ type ChainFile struct {
 func (cf *ChainFile) ToInitializationParameters() primitives.InitializationParameters {
 	ip := primitives.InitializationParameters{
 		InitialValidators: cf.Validators,
-		GenesisTime:       time.Unix(cf.GenesisTime, 0),
+		GenesisTime:       time.Unix(int64(cf.GenesisTime), 0),
 		PremineAddress:    cf.PremineAddress,
 	}
